backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3.0    # for JWT RS256 verification + JWKS handling
requests>=2.31
django-environ>=0.10.0
django-cors-headers>=3.14
django-ratelimit>=3.0
django-prometheus>=2.0
sentry-sdk>=1.20
celery>=5.3
redis>=4.5
boto3>=1.26        # if using S3
django-storages>=1.14  # optional for S3 backend
pytest>=7.2
pytest-django>=4.5
factory-boy>=3.2
drf-spectacular>=0.28   # OpenAPI / schema generation
